(feat) add helper fct for pkg-config
